+++ spec/controllers/api/v1/games_controller_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-describe Api::V1::OraculsController do
+describe Api::V1::GamesController do
   describe 'GET#index' do
     it_behaves_like 'required api auth'
     it_behaves_like 'required api email confirmation'
@@ -8,25 +8,22 @@ describe Api::V1::OraculsController do
 
     context 'for logged users' do
       let!(:user) { create :user }
+      let!(:week) { create :week }
       let(:access_token) { Auth::GenerateTokenService.new.call(user: user)[:result] }
 
-      before do
-        create_list :oracul, 2
-        create :oracul, user: user
-      end
+      before { create_list :game, 2, week: week }
 
-      it 'returns oracul data', :aggregate_failures do
+      it 'returns games data', :aggregate_failures do
         get :index, params: {
-          api_access_token: access_token, response_include_fields: 'name,oracul_place_id'
+          api_access_token: access_token, response_include_fields: 'id,points,predictable'
         }
 
         expect(response).to have_http_status :ok
 
-        attributes = response.parsed_body.dig('oraculs', 'data', 0, 'attributes')
-        expect(response.parsed_body.dig('oraculs', 'data').size).to eq 1
-        expect(attributes['id']).to be_nil
-        expect(attributes['name']).not_to be_nil
-        expect(attributes['oracul_place_id']).not_to be_nil
+        attributes = response.parsed_body.dig('games', 'data', 0, 'attributes')
+        expect(attributes['id']).not_to be_nil
+        expect(attributes['points']).not_to be_nil
+        expect(attributes['predictable']).not_to be_nil
       end
     end