+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - url localization
 - api v1 endpoints for user login
+- api v1 endpoints for forecasts
 
 ### Modified
 - user navigation